Not requiring AUFS
This is more of a thought experiment.

But adding `|| true` to the line that attempts to install `linux-image-extra-...` would mean that dokku could run places with custom kernels, like Linode, by falling back to DeviceMapper. People who really want AUFS could use pv-grub and the StackScript (once a documentation tweak is made).

Is that just a crazy idea? I know that DeviceMapper has worse performance for starting/stopping containers, but as AUFS isn't upstream, it seems like it might work ...